Extract impl matching logic
