+++ crates/engine/body-ir/src/resolution/mod.rs
@@ -5,6 +5,7 @@
 
 mod autoderef;
 mod body;
+mod impl_match;
 mod index;
 mod method;
 mod normalize;
@@ -27,11 +28,11 @@ use crate::{
 
 use self::{
     body::BodyValuePathResolver,
+    impl_match::BodyImplMatcher,
     method::{
         local_function_applies_to_receiver as local_function_applies_to_receiver_impl,
         semantic_function_applies_to_receiver as semantic_function_applies_to_receiver_impl,
         semantic_trait_function_candidates_for_receiver as semantic_trait_function_candidates_for_receiver_impl,
-        semantic_trait_impl_applicability,
     },
     ty::{TypeSubst, ty_from_type_ref_in_context},
     type_path::BodyTypePathResolver,
@@ -123,10 +124,9 @@ pub(super) fn semantic_trait_impl_applies_to_receiver(
     trait_impl: TraitImplRef,
     receiver_ty: &BodyNominalTy,
 ) -> Result<bool, PackageStoreError> {
-    Ok(
-        semantic_trait_impl_applicability(def_map, semantic_ir, trait_impl, receiver_ty)?
-            .is_applicable(),
-    )
+    Ok(BodyImplMatcher::new(def_map, semantic_ir)
+        .semantic_trait_impl_applicability(trait_impl, receiver_ty)?
+        .is_applicable())
 }
 
 pub(super) fn local_function_applies_to_receiver(
